ARM: dts: Use haptic node from qcom,qpnp-haptic
This needs to be change, because the haptic import from Asus
is in drivers/soc/qcom/qpnp-haptic.c file and it uses the node
"qcom,qpnp-haptic".

Meanwhile, the latest CAF changes use the node "qcom,qpnp-haptics"
which comes from the drivers/leds/leds-qpnp-haptics.c file and it is
enabled by default on sdm660_defconfig file.

I am assuming, if this is not changed it is possible that the
haptic driver will not work or cannot be controlled.

It also adapts to Import from Asus changes.